Rename createflag
