+++ frontend/web/components/modals/create-feature/tabs/FeatureSettings.tsx
