+++ frontend/web/components/modals/create-feature/tabs/FeatureValue.tsx
